fix(api): enforce folder access check in unlinkFolder endpoint

Tested-by: shaheem.azmal@siemens.com

+++ src/www/ui/api/Controllers/FolderController.php
@@ -310,7 +310,11 @@ class FolderController extends RestController
       throw new HttpNotFoundException("Folder content id not found!");
     }
     /** @var FolderDao $folderDao */
-    $folderDao = $this->container->get('dao.folder');
+    $folderDao = $this->restHelper->getFolderDao();
+    $content = $folderDao->getContent($folderContentId);
+    if (! $folderDao->isFolderAccessible($content['parent_fk'], $this->restHelper->getUserId())) {
+      throw new HttpForbiddenException("Folder is not accessible!");
+    }
     if (!$folderDao->removeContent($folderContentId)) {
       throw new HttpBadRequestException("Content cannot be unlinked.");
     }
